backend/requirements.txt

email-validator
aiofiles
typing-extensions
alembic

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
boto3==1.34.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# เพิ่มตามที่โค้ดใช้
slowapi==0.1.9
limits==5.5.0
requests>=2.32
httpx==0.28.1
Authlib>=1.6
PyJWT>=2.8
pydantic[email]==2.5.0
itsdangerous>=2.1

passlib[bcrypt]

# MediaPipe and Computer Vision
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.21.0